Add cache helper to save break reason list that not clear after finish process
